pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "geomosaic"
version = "1.3.0"
authors = [
  { name = "Davide Corso", email = "davidecrs92@gmail.com" },
  { name = "Edoardo Taccaliti", email = "edoardotaccaliti@gmail.com" },
  { name = "Guillermo Climent Gargallo", email = "guillermo.climentgargallo@unina.it" },
]
description = "A flexible metagenomic pipeline combining read-based, assemblies and MAGs with downstream analysis"
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3"
]
license = "GPL-3.0+"
license-files = ["LICENSE"]
dependencies = [
    "pandas",
    "networkx",
    "pyyaml",
    "matplotlib",
    "snakemake",
    "biopython>=1.82",
    "tqdm",
    "openpyxl",
    "pulp<2.8"
]

[project.scripts]
geomosaic = "geomosaic.geomosaic:main"

[project.urls]
"Homepage" = "https://github.com/giovannellilab/Geomosaic"
"Bug Tracker" = "https://github.com/giovannellilab/Geomosaic/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
geomosaic = ["*.json"]
"geomosaic.envs" = ["*.yaml"]
"geomosaic.modules" = ["**/*.smk", "**/*.yaml"]
"geomosaic.modules_extdb" = ["**/*.smk", "**/*.txt"]
